64-bit warning fixes
git-svn-id: https://svn.wxwidgets.org/svn/wx/wxWidgets/trunk@41547 c3d73ce0-8a6f-49c7-b76d-6d57e0e08775

+++ samples/event/event.cpp
@@ -97,7 +97,7 @@ public:
 
 protected:
     // number of pushed event handlers
-    size_t m_nPush;
+    unsigned m_nPush;
 
 private:
     // any class wishing to process wxWidgets events must use this macro
@@ -120,7 +120,7 @@ public:
     }
 
 private:
-    size_t m_level;
+    unsigned m_level;
 
     DECLARE_EVENT_TABLE()
 };
